requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
chromadb
tiktoken
unstructured[pdf,docx,html]
python-dotenv
pydantic
